+++ src/matterbridge.ts
@@ -476,8 +476,8 @@ export class Matterbridge extends EventEmitter {
         // Update the plugin information
         plugin.name = packageJson.name as string;
         plugin.version = packageJson.version as string;
-        plugin.description = packageJson.description as string;
-        plugin.author = packageJson.author as string;
+        plugin.description = (packageJson.description as string) ?? 'No description';
+        plugin.author = (packageJson.author as string) ?? 'Unknown';
       } else {
         this.log.info(`Error parsing plugin ${plg}${plugin.name}${nf}. Trying to reinstall it from npm.`);
         try {
@@ -1723,9 +1723,9 @@ export class Matterbridge extends EventEmitter {
         platform.config = config;
         platform.version = packageJson.version;
         plugin.name = packageJson.name;
-        plugin.description = packageJson.description;
+        plugin.description = packageJson.description ?? 'No description';
         plugin.version = packageJson.version;
-        plugin.author = packageJson.author;
+        plugin.author = packageJson.author ?? 'Unknown';
         plugin.type = platform.type;
         plugin.platform = platform;
         plugin.loaded = true;